requirements.txt
requests
beautifulsoup4
#tavily
groq
jinja2
apscheduler
python-dotenv